Refactoring scripts. See comments.
