+++ Assets/Scripts/Movement.cs
@@ -1,13 +1,12 @@
 using UnityEngine;
 
-[RequireComponent(
-    typeof(Rigidbody2D))]
+[RequireComponent(typeof(Rigidbody2D))]
 
 public class Movement : MonoBehaviour
 {
+    private const string Horizontal = "Horizontal";
+    
     [SerializeField] private float _speed = 5.0f;
-
-    private const string _horizontal = "Horizontal";
     
     private Rigidbody2D _rigidbody2D;
     private float _horizontalMove;
@@ -17,7 +16,7 @@ public class Movement : MonoBehaviour
 
     private void Update()
     {
-        _horizontalMove = Input.GetAxis(_horizontal) * _speed;
+        _horizontalMove = Input.GetAxis(Horizontal) * _speed;
         _rigidbody2D.velocity = new Vector2(_horizontalMove, _rigidbody2D.velocity.y);
     }
 }
